backend/pyproject.toml
[tool.poetry]
name = "leetconnect-backend"
version = "0.1.0"
description = ""
authors = ["Kailash Choudhary <kailashjat604@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "3.2.5"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
